+++ src/ehrdata/core/index.py
@@ -3,11 +3,7 @@ from functools import singledispatch
 
 import numpy as np
 import pandas as pd
-
-try:  # anndata 0.13: aliases moved to anndata.typing
-    from anndata.typing import Index
-except ImportError:
-    from anndata.compat import Index
+from anndata.typing import Index
 
 
 @singledispatch